+++ robotcode/language_server/robotframework/parts/document_highlight.py
@@ -174,6 +174,8 @@ class RobotDocumentHighlightProtocolPart(RobotLanguageServerProtocolPart, ModelH
         if result is not None:
             keyword_doc, keyword_token = result
 
+            keyword_token = self.strip_bdd_prefix(keyword_token)
+
             lib_entry, kw_namespace = await self.get_namespace_info_from_keyword(namespace, keyword_token)
 
             kw_range = range_from_token(keyword_token)
@@ -185,7 +187,12 @@ class RobotDocumentHighlightProtocolPart(RobotLanguageServerProtocolPart, ModelH
                 if position in r:
                     # TODO highlight namespaces
                     return None
-            if keyword_doc is not None and not keyword_doc.is_error_handler and keyword_doc.source:
+            if (
+                position in kw_range
+                and keyword_doc is not None
+                and not keyword_doc.is_error_handler
+                and keyword_doc.source
+            ):
                 return [
                     *(
                         [DocumentHighlight(keyword_doc.range, DocumentHighlightKind.TEXT)]
@@ -258,6 +265,9 @@ class RobotDocumentHighlightProtocolPart(RobotLanguageServerProtocolPart, ModelH
 
         if result is not None:
             keyword_doc, keyword_token = result
+
+            keyword_token = self.strip_bdd_prefix(keyword_token)
+
             lib_entry, kw_namespace = await self.get_namespace_info_from_keyword(namespace, keyword_token)
 
             kw_range = range_from_token(keyword_token)
@@ -270,7 +280,7 @@ class RobotDocumentHighlightProtocolPart(RobotLanguageServerProtocolPart, ModelH
                     # TODO highlight namespaces
                     return None
 
-            if keyword_doc is not None and not keyword_doc.is_error_handler:
+            if position in kw_range and keyword_doc is not None and not keyword_doc.is_error_handler:
                 return [
                     *(
                         [DocumentHighlight(keyword_doc.range, DocumentHighlightKind.TEXT)]
@@ -299,6 +309,8 @@ class RobotDocumentHighlightProtocolPart(RobotLanguageServerProtocolPart, ModelH
             keyword_token = cast(RobotToken, template_node.get_token(RobotToken.NAME))
             if keyword_token is None:
                 return None
+
+            keyword_token = self.strip_bdd_prefix(keyword_token)
 
             if position.is_in_range(range_from_token(keyword_token)):
                 namespace = await self.parent.documents_cache.get_namespace(document)